% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sentences.r
\name{generate_sentences}
\alias{generate_sentences}
\title{Generate sentences}
\usage{
generate_sentences(
  seu_obj,
  sen_len = 260,
  region_info,
  save_inseu = TRUE,
  save_path,
  genes_representor,
  envir_path
)
}
\arguments{
\item{seu_obj}{The seurat object. The gene expression counts matrix will be extracted from this object.}

\item{sen_len}{Default is 260. The length of gene word sentence, the generated sentences length will be minus or equal this parameter. Genes with zero expression levels will not be included in the sentences.}

\item{region_info}{The region information of TLS. If there lack or not contain region information, this parameter can be set as NULL. The region information 1 is represeted tha target region, 0 is represeted the other region}

\item{save_inseu}{Default is TRUE. Whether save sentences in the seurat meta.data. If set FALSE, the sentences will be save in a .txt file, which include the sentences and meta.data information.}

\item{save_path}{If the save_inseu was setted as FALSE, the save path must be provided, this parameter is the sentences storage file path.}

\item{genes_representor}{The file path of genes were used in the pre-training the gene word encoder.}

\item{envir_path}{The python env path.}
}
\value{
The sentences of single cells or spots which are saved in the meta.data of seurat object or saved in the single txt file.
}
\description{
Based on the expression levels of each gene to generate sentence of each single cell or spot.
}
